@@ -7,14 +7,14 @@ plugins {
 
 android {
     namespace = "se.curtrune.lucy"
-    compileSdk = 35
+    compileSdk = 36
 
     defaultConfig {
         applicationId = "se.curtrune.lucy"
         minSdk = 24
         targetSdk = 34
-        versionCode = 18
-        versionName = "creature of doom"
+        versionCode = 19
+        versionName = "its the truth"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
         vectorDrawables {
