@@ -3,5 +3,5 @@ package se.curtrune.lucy.app;
 import java.io.Serializable;
 
 public enum InitialScreen implements Serializable {
-    TODO_FRAGMENT, CALENDER_DATE, CALENDER_WEEK, CALENDER_MONTH, CALENDER_APPOINTMENTS
+    TODO_FRAGMENT, CALENDER_DATE, CALENDER_WEEK, CALENDER_MONTH, CALENDER_APPOINTMENTS, NEW_DAY_CALENDER
 }
