@@ -3,9 +3,12 @@ package se.curtrune.lucy.screens.dev.composables
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.Button
 import androidx.compose.material3.Card
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
@@ -21,6 +24,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
 import se.curtrune.lucy.R
 import se.curtrune.lucy.screens.duration.DurationEvent
 import se.curtrune.lucy.screens.duration.DurationState
@@ -76,7 +80,8 @@ fun DurationControls(state: DurationState, onEvent: (DurationEvent)->Unit){
                 mutableIntStateOf(0)
             }
             Row(modifier = Modifier.fillMaxWidth()){
-                val options = listOf("category", "date", "tags")
+                //val options = listOf("category", "date", "tags")
+                val options = listOf("category", "date")
                 SingleChoiceSegmentedButtonRow {
                     options.forEachIndexed{ index, label->
                         SegmentedButton(onClick = {
@@ -105,7 +110,14 @@ fun DurationControls(state: DurationState, onEvent: (DurationEvent)->Unit){
                     }
                 }
             }
+            Spacer(modifier = Modifier.height(8.dp))
             Text(text = "total duration: ${DateTImeConverter.formatSeconds(state.totalDuration)}")
+            Spacer(modifier = Modifier.height(8.dp))
+            Button(onClick = {
+                onEvent(DurationEvent.Refresh)
+            }){
+                Text(text = "refresh")
+            }
         }
     }
 }
