@@ -18,6 +18,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import se.curtrune.lucy.app.InitialScreen
 import se.curtrune.lucy.screens.log_in.ui.theme.LucyTheme
 import se.curtrune.lucy.screens.main.MainActivity
+import se.curtrune.lucy.screens.main.MainActivity2
 import se.curtrune.lucy.util.Constants
 import se.curtrune.lucy.util.Logger.Companion.log
 
@@ -54,14 +55,18 @@ class LogInActivity : ComponentActivity() {
         }
     }
     private fun startUserActivity(initialScreen: InitialScreen) {
-        log("...startUserActivity()")
-        val intent=Intent(applicationContext,MainActivity::class.java)
+        println("LogInActivity.startUserActivity()")
+        var intent=Intent(applicationContext, MainActivity::class.java)
         when(initialScreen){
             InitialScreen.TODO_FRAGMENT-> {intent.putExtra(Constants.INITIAL_SCREEN, initialScreen.name)}
             InitialScreen.CALENDER_DATE -> {intent.putExtra(Constants.INITIAL_SCREEN,initialScreen.name)}
             InitialScreen.CALENDER_WEEK -> {intent.putExtra(Constants.INITIAL_SCREEN,initialScreen.name)}
             InitialScreen.CALENDER_MONTH -> {intent.putExtra(Constants.INITIAL_SCREEN,initialScreen.name)}
             InitialScreen.CALENDER_APPOINTMENTS ->{intent.putExtra(Constants.INITIAL_SCREEN,initialScreen.name)}
+            InitialScreen.NEW_DAY_CALENDER -> {
+                intent = Intent(applicationContext, MainActivity2::class.java)
+                intent.putExtra(Constants.INITIAL_SCREEN,initialScreen.name)
+            }
         }
         startActivity(intent)
     }
