@@ -10,7 +10,7 @@ sealed interface MainChannelEvent{
     data class ShowAffirmation(val affirmation: Affirmation? = null): MainChannelEvent
     data class ShowMessage(val message: String): MainChannelEvent
     data object ShowPanicDialog: MainChannelEvent
-    data object OpenNavigationDrawer : MainChannelEvent
+    data class ShowNavigationDrawer(val show: Boolean) : MainChannelEvent
     data object ShowDayCalendar : MainChannelEvent
     data class StartSequence(val root: Item) : MainChannelEvent
     data class ShowQuoteDialog(val quote: Quote? = null): MainChannelEvent
