@@ -26,7 +26,6 @@ import se.curtrune.lucy.R
 import se.curtrune.lucy.activities.kotlin.ui.theme.LucyTheme
 import se.curtrune.lucy.classes.item.Item
 import se.curtrune.lucy.composables.top_app_bar.TopAppBarEvent
-import se.curtrune.lucy.screens.item_editor.ItemEditorFragment
 import se.curtrune.lucy.screens.main.MainViewModel
 import se.curtrune.lucy.screens.medicine.composable.AddMedicineFab
 import se.curtrune.lucy.screens.medicine.composable.AdverseEffectDialog
@@ -118,11 +117,12 @@ class MedicineFragment : Fragment() {
     private fun navigateToEditor(item: Item) {
         println("...navigateToEditor(${item.heading})")
         val mainViewModel = ViewModelProvider(requireActivity())[MainViewModel::class.java]
-        mainViewModel.updateFragment(
+        throw Exception("not implemented, deprecated")
+/*        mainViewModel.updateFragment(
             ItemEditorFragment(
                 item
             )
-        )
+        )*/
     }
 }
 
