@@ -3,5 +3,5 @@ package se.curtrune.lucy.screens.monthcalendar
 sealed interface MonthChannel{
     data class ShowMessage(val message: String): MonthChannel
     data object ShowAddItemDialog: MonthChannel
-    data object NavigateToDayCalendar: MonthChannel
+    data class NavigateToDayCalendar(val date: String): MonthChannel
 }
